Add files via upload
This is a three minute thesis that was created for my GIS Applications class. The question is was there an adequate number of hospitals on the Middle Peninsula for the number of vehicular accidents with injuries. The answer was no. Based on a 7 to 12 minute travel time for ambulances from the scene of the accident to the area hospitals, one area in the southern Newport News area of Virginia needed an additional hospital.