Add files via upload
This is a map of one of the southwestern Japan Islands that was created for my job. There are 25 separate shapefiles for various features of the island. The features range from roads and buildings, to vegetation and water features of all types, to airport features and communication towers. This will then be uploaded into a 3D software to be turned into a terrain based simulation for the military,