Add files via upload
This map is a new bypass that I proposed in my final paper for my Bachelor's Degree. The route in red runs from just south of Baltimore, MD, to south of Quantico, VA and up to southeast of Frederick, MD. This proposed bypass was suggested to reroute the heavier traffic away from Northern Virginia.